refactor: 엔티티에서 DTO 의존성 제거 및 도메인 서비스(RouteAdviceService) 도입
