+++ apps/server/src/main/java/com/weanet/server/domain/RouteStep.java
@@ -81,19 +81,4 @@ public class RouteStep extends BaseEntity {
     public void assignRoute(Route route) {
         this.route = route;
     }
-
-    public RouteStepResponse toResponse() {
-        return RouteStepResponse.builder()
-                .sequence(this.sequence)
-                .transportType(this.transportType)
-                .lineName(this.lineName)
-                .lineId(this.lineId)
-                .startStationName(this.startStationName)
-                .startStationId(this.startStationId)
-                .endStationName(this.endStationName)
-                .endStationId(this.endStationId)
-                .lat(this.lat)
-                .lng(this.lng)
-                .build();
-    }
 }
