+++ apps/server/src/main/java/com/weanet/server/service/DashboardService.java
@@ -21,6 +21,7 @@ public class DashboardService {
     private final RouteRepository routeRepository;
     private final ReportRepository reportRepository;
     private final RouteEnrichmentService enrichmentService;
+    private final RouteAdviceService adviceService;
 
     /**
      * 특정 경로의 실시간 데이터를 종합하여 통합 리포트를 생성합니다.
@@ -29,17 +30,28 @@ public class DashboardService {
         Route route = routeRepository.findById(routeId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 경로를 찾을 수 없습니다."));
 
-        // 1. 구간별 실시간 정보 보강을 위한 DTO 변환 (엔티티 메소드 사용)
+        // 1. 구간별 실시간 정보 보강을 위한 DTO 변환 (서비스 레이어에서 수행)
         List<RouteStepResponse> steps = route.getSteps().stream()
-                .map(com.weanet.server.domain.RouteStep::toResponse)
+                .map(step -> RouteStepResponse.builder()
+                        .sequence(step.getSequence())
+                        .transportType(step.getTransportType())
+                        .lineName(step.getLineName())
+                        .lineId(step.getLineId())
+                        .startStationName(step.getStartStationName())
+                        .startStationId(step.getStartStationId())
+                        .endStationName(step.getEndStationName())
+                        .endStationId(step.getEndStationId())
+                        .lat(step.getLat())
+                        .lng(step.getLng())
+                        .build())
                 .collect(Collectors.toList());
 
         // 2. 실시간 데이터 보강
         enrichmentService.enrichRoute(steps);
 
-        // 3. 통합 조언 및 요약 정보 생성 (엔티티 내 비즈니스 로직 사용)
-        String integratedAdvice = Route.generateAdvice(steps);
-        List<RouteIntegratedReportResponse.StepSummary> stepSummaries = route.getStepSummaries(steps);
+        // 3. 통합 조언 및 요약 정보 생성 (도메인 서비스 사용)
+        String integratedAdvice = adviceService.generateAdvice(steps);
+        List<RouteIntegratedReportResponse.StepSummary> stepSummaries = adviceService.getStepSummaries(steps);
 
         // 4. 해당 노선들에 대한 최신 제보 조회
         List<ReportResponse> recentReports = reportRepository.findByRouteIdOrderByCreatedAtDesc(routeId).stream()
